requirements.txt
telethon==1.36.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
